feat(auth): inner-circle tier — /curation surfaces for invited adjudicators (#3846)
Derek needs to hand the identity-review queues to an outside collaborator
without granting superadmin, and the main site has no middle tier (roles
resolve to superadmin-or-reader; inner_circle docs were dead).

- requireInnerCircle / withInnerCircleAuth: superadmin OR a global
  memberships doc {tenantId: null, role: inner_circle|editor, active}.
  Checked live against the DB (grant applies without re-login, #3048
  rationale). Deliberately NOT minted into the session role — that would
  unlock every editor-gated surface (imports, metadata edits) for someone
  invited only to adjudicate. Deliberately ignores the x-tenant-id
  escalation — a tenant editor must not reach global actuation through a
  subdomain session. No CRON bypass: nothing unattended may actuate these
  queues.
- Identity review moves /admin/identity-review → /curation/identity-review
  (old path redirects; AdminNav updated). /curation/* gates on the new
  tier; both API routes switch withAdminAuth → withInnerCircleAuth.
- /curation joins tenant-global-paths: corpus-wide and actuating, so
  partner subdomains refuse the prefix outright.
- The legacy requireInnerCircle/withInnerCircleAuth aliases (mapped to
  requireRole('editor'), which never resolves on the main site — i.e.
  superadmin-only in practice) are removed; their four callsites
  (/analytics, /experiments, /jobs, /traffic) now open to inner-circle
  members, which is what the alias name always promised. Flagged as a
  conscious scope decision.
