+++ src/app/admin/AdminNav.tsx
@@ -14,7 +14,7 @@ interface NavItem {
 const adminLinks: NavItem[] = [
   { href: '/admin', label: 'Dashboard', exact: true },
   { href: '/admin/canon', label: 'Canon' },
-  { href: '/admin/identity-review', label: 'Identity review' },
+  { href: '/curation/identity-review', label: 'Identity review' },
   { href: '/admin/pipeline', label: 'Pipeline' },
   { href: '/admin/processing', label: 'Processing' },
   { href: '/admin/realtime', label: 'Realtime' },
